Refactored state util funcs into their own file
